merger run_impl and run
