+++ olive/workflows/__init__.py
@@ -2,6 +2,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-from olive.workflows.run_impl import run
+from olive.workflows.run import run
 
 __all__ = ["run"]
